Code reorganized
