+++ designer/client/pages/landingpage/ChooseExisting.tsx
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
-import * as formConfigurationApi from "../load-form-configurations";
-import { withI18n } from "../i18n";
+import * as formConfigurationApi from "../../load-form-configurations";
+import { withI18n } from "../../i18n";
 import { withRouter } from "react-router-dom";
-import { BackLink } from "../components/BackLink";
+import { BackLink } from "../../components/BackLink";
 import "./LandingPage.scss";
 
 type Props = {
