+++ designer/client/pages/landingpage/__tests__/Choice.jest.tsx
